#123: Flyttar alle repo-klassane inn i tilhøyrande modellklasse

+++ src/main/kotlin/no/roedt/ringesentralen/dashboard/DashboardService.kt
@@ -26,9 +26,9 @@ class DashboardServiceBean(
 
         val statusliste: List<Lokallagsstatus> = mineLokallag
                 .map { lokallag ->
-                    val igjenAaRinge = personRepository.list("groupID<2 and lokallag=?1", lokallag.id).size
+                    val igjenAaRinge = personRepository.list("groupID<2 and lokallag=?1", lokallag.id.toInt()).size
                     val personerSomKanRinges = databaseUpdater.getResultList("SELECT 1 FROM v_personerSomKanRinges WHERE lokallag = ${lokallag.id}").size
-                    val totaltInklRingte = personRepository.list("lokallag=?1 and groupID=2", lokallag.id).size
+                    val totaltInklRingte = personRepository.list("lokallag=?1 and groupID=2", lokallag.id.toInt()).size
                     Lokallagsstatus(
                             lokallag = lokallag,
                             igjenAaRinge = igjenAaRinge,
